Time: 3 ms (35.83%), Space: 6 MB (26.76%) - LeetHub

+++ 371-sum-of-two-integers/371-sum-of-two-integers.cpp
@@ -4,9 +4,9 @@ public:
     {
         while(b)
         {
-            unsigned int carry = a & b;
+            int carry = a & b;
             a = a ^ b;
-            b = carry  << 1;
+            b = (unsigned int)carry  << 1;
         }
         return a;
     }
